+++ bindings/python/pymongoarrow/api.py
@@ -51,7 +51,9 @@ from pyarrow.types import (
 )
 from pymongo.collection import Collection
 from pymongo.common import MAX_WRITE_BATCH_SIZE
+from pymongo.driver_info import DriverInfo
 
+import pymongoarrow.version as pymongoarrow_version
 from pymongoarrow.context import PyMongoArrowContext
 from pymongoarrow.errors import ArrowWriteError
 from pymongoarrow.result import ArrowWriteResult
@@ -90,6 +92,15 @@ _MAX_MESSAGE_SIZE = 48000000 - 16 * 1024
 _MAX_WRITE_BATCH_SIZE = max(100000, MAX_WRITE_BATCH_SIZE)
 
 
+def _add_driver_metadata(collection: Collection):
+    client = collection.database.client
+
+    if callable(client.append_metadata):
+        client.append_metadata(
+            DriverInfo(name="PyMongoArrow", version=pymongoarrow_version.__version__)
+        )
+
+
 def find_arrow_all(collection, query, *, schema=None, allow_invalid=False, **kwargs):
     """Method that returns the results of a find query as a
     :class:`pyarrow.Table` instance.
@@ -110,6 +121,7 @@ def find_arrow_all(collection, query, *, schema=None, allow_invalid=False, **kwa
     :Returns:
       An instance of class:`pyarrow.Table`.
     """
+    _add_driver_metadata(collection)
     context = PyMongoArrowContext(
         schema, codec_options=collection.codec_options, allow_invalid=allow_invalid
     )
@@ -152,6 +164,7 @@ def aggregate_arrow_all(collection, pipeline, *, schema=None, allow_invalid=Fals
     :Returns:
       An instance of class:`pyarrow.Table`.
     """
+    _add_driver_metadata(collection)
     context = PyMongoArrowContext(
         schema, codec_options=collection.codec_options, allow_invalid=allow_invalid
     )
@@ -340,39 +353,13 @@ def aggregate_numpy_all(collection, pipeline, *, schema=None, allow_invalid=Fals
     )
 
 
-def _cast_away_extension_type(field: pa.field) -> pa.field:
-    if isinstance(field.type, pa.ExtensionType):
-        field_without_extension = pa.field(field.name, field.type.storage_type)
-    elif isinstance(field.type, pa.StructType):
-        field_without_extension = pa.field(
-            field.name,
-            pa.struct([_cast_away_extension_type(nested_field) for nested_field in field.type]),
-        )
-    elif isinstance(field.type, pa.ListType):
-        field_without_extension = pa.field(
-            field.name, pa.list_(_cast_away_extension_type(field.type.value_field))
-        )
-    else:
-        field_without_extension = field
-
-    return field_without_extension
-
-
 def _arrow_to_polars(arrow_table: pa.Table):
-    """Helper function that converts an Arrow Table to a Polars DataFrame.
-
-    Note: Polars lacks ExtensionTypes. We cast them  to their base arrow classes.
-    """
+    """Helper function that converts an Arrow Table to a Polars DataFrame."""
     if pl is None:
         msg = "polars is not installed. Try pip install polars."
         raise ValueError(msg)
 
-    schema_without_extensions = pa.schema(
-        [_cast_away_extension_type(field) for field in arrow_table.schema]
-    )
-    arrow_table_without_extensions = arrow_table.cast(schema_without_extensions)
-
-    return pl.from_arrow(arrow_table_without_extensions)
+    return pl.from_arrow(arrow_table)
 
 
 def find_polars_all(collection, query, *, schema=None, allow_invalid=False, **kwargs):
@@ -570,6 +557,7 @@ def write(
     else:
         type_registry = TypeRegistry([*base_codecs, _DecimalCodec()])
     codec_options = codec_options.with_options(type_registry=type_registry)
+    _add_driver_metadata(collection)
 
     while cur_offset < tab_size:
         cur_size = 0