+++ bindings/python/test/pandas_types/test_decimal128.py
@@ -11,6 +11,8 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+from test.pandas_types.util import base_make_data
+
 import numpy as np
 import pytest
 from bson import Decimal128
@@ -40,13 +42,7 @@ def dtype():
 
 
 def make_data():
-    return (
-        [make_datum() for _ in range(8)]
-        + [np.nan]
-        + [make_datum() for _ in range(88)]
-        + [np.nan]
-        + [make_datum(), make_datum()]
-    )
+    return base_make_data(make_datum)
 
 
 @pytest.fixture